refactor: move stateid to separate file
