+++ types/block.go
@@ -6,7 +6,6 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
-	"strconv"
 	"strings"
 	"time"
 
@@ -711,51 +710,6 @@ func (commit *Commit) CanonicalVoteVerifySignID(chainID string, quorumType btcjs
 	return VoteBlockSignID(chainID, vCanonical, quorumType, quorumHash)
 }
 
-// VoteStateSignID returns the signID bytes of the state for the Vote corresponding to valIdx for
-// signing.
-//
-// Panics if valIdx >= commit.Size().
-//
-func (commit *Commit) VoteStateSignID(chainID string, quorumType btcjson.LLMQType, quorumHash []byte) []byte {
-	stateID := commit.StateID.ToProto()
-
-	return VoteStateSignID(chainID, stateID, quorumType, quorumHash)
-}
-
-// VoteStateRequestId returns the requestId Hash of the Vote corresponding to valIdx for
-// signing.
-//
-// Panics if valIdx >= commit.Size().
-//
-func (commit *Commit) VoteStateRequestID() []byte {
-	requestIDMessage := []byte("dpsvote")
-	heightByteArray := make([]byte, 8)
-	binary.LittleEndian.PutUint64(heightByteArray, uint64(commit.Height))
-	roundByteArray := make([]byte, 4)
-	binary.LittleEndian.PutUint32(roundByteArray, uint32(commit.Round))
-
-	requestIDMessage = append(requestIDMessage, heightByteArray...)
-	requestIDMessage = append(requestIDMessage, roundByteArray...)
-
-	return crypto.Sha256(requestIDMessage)
-}
-
-// CanonicalVoteStateSignBytes returns the bytes of the State corresponding to valIdx for
-// signing.
-//
-// Panics if valIdx >= commit.Size().
-//
-// See VoteSignBytes
-func (commit *Commit) CanonicalVoteStateSignBytes(chainID string) []byte {
-	stateID := commit.StateID.ToProto()
-	return VoteStateSignBytes(chainID, stateID)
-}
-
-func (commit *Commit) CanonicalVoteStateSignID(chainID string, quorumType btcjson.LLMQType, quorumHash []byte) []byte {
-	stateID := commit.StateID.ToProto()
-	return VoteStateSignID(chainID, stateID, quorumType, quorumHash)
-}
-
 // Type returns the vote type of the commit, which is always VoteTypePrecommit
 // Implements VoteSetReader.
 func (commit *Commit) Type() byte {
@@ -1185,95 +1139,4 @@ func BlockIDFromProto(bID *tmproto.BlockID) (*BlockID, error) {
 	blockID.Hash = bID.Hash
 
 	return blockID, blockID.ValidateBasic()
-}
-
-//--------------------------------------------------------------------------------
-
-// StateID
-// TODO: Move to separate file
-type StateID struct {
-
-	// Height of last committed block
-	Height int64 `json:"height"`
-	// LastAppHash used in last committed block
-	LastAppHash tmbytes.HexBytes `json:"last_app_hash"`
-}
-
-// Copy returns new StateID that is equal to this one
-func (stateID StateID) Copy() StateID {
-	appHash := make([]byte, len(stateID.LastAppHash))
-	if copy(appHash, stateID.LastAppHash) != len(stateID.LastAppHash) {
-		panic("Cannot copy LastAppHash, this should never happen. Out of memory???")
-	}
-
-	return StateID{
-		Height:      stateID.Height,
-		LastAppHash: appHash,
-	}
-}
-
-// Equals returns true if the StateID matches the given StateID
-func (stateID StateID) Equals(other StateID) bool {
-	return stateID.Height == other.Height &&
-		bytes.Equal(stateID.LastAppHash, other.LastAppHash)
-}
-
-// Key returns a machine-readable string representation of the StateID
-func (stateID StateID) Key() string {
-	return strconv.FormatInt(stateID.Height, 36) + string(stateID.LastAppHash)
-}
-
-// ValidateBasic performs basic validation.
-func (stateID StateID) ValidateBasic() error {
-	// LastAppHash can be empty in case of genesis block.
-	if err := ValidateAppHash(stateID.LastAppHash); err != nil {
-		return fmt.Errorf("wrong app Hash")
-	}
-
-	if stateID.Height < 0 {
-		return fmt.Errorf("stateID height is not valid: %d < 0", stateID.Height)
-	}
-
-	return nil
-}
-
-// String returns a human readable string representation of the StateID.
-//
-// 1. hash
-//
-func (stateID StateID) String() string {
-	return fmt.Sprintf(`%d:%v`, stateID.Height, stateID.LastAppHash)
-}
-
-// ToProto converts BlockID to protobuf
-func (stateID StateID) ToProto() tmproto.StateID {
-	return tmproto.StateID{
-		LastAppHash: stateID.LastAppHash,
-		Height:      stateID.Height,
-	}
-}
-
-// WithHeight returns new copy of stateID with height set to provided value.
-// It is a convenience method used in tests.
-// Note that this is Last Height from state, so it will be (height-1) for Vote.
-func (stateID StateID) WithHeight(height int64) StateID {
-	ret := stateID.Copy()
-	ret.Height = height
-
-	return ret
-}
-
-// FromProto sets a protobuf BlockID to the given pointer.
-// It returns an error if the block id is invalid.
-func StateIDFromProto(sID *tmproto.StateID) (*StateID, error) {
-	if sID == nil {
-		return nil, errors.New("nil StateID")
-	}
-
-	stateID := new(StateID)
-
-	stateID.LastAppHash = sID.LastAppHash
-	stateID.Height = sID.Height
-
-	return stateID, stateID.ValidateBasic()
 }
